# swagger.yaml
openapi: 3.0.0
info:
  title: Jobs API
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:3000/api/v1
    variables: {}
paths:
...
  /jobs/{id}:
    patch:
      tags:
        - Jobs
      summary: Update Job
      operationId: UpdateJob
      parameters: [id]
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UpdateJobRequest'
                - example:
                    company: apple
            example:
              company: apple
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
        - bearer: []
    delete:
      tags:
        - Jobs
      summary: Delete Job
      operationId: DeleteJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
...
    UpdateJobRequest:
      title: UpdateJobRequest
      required:
        - company
      type: object
      properties:
        company:
          type: string
      example:
        company: apple
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...